Decouple Lit UI component catalog from the core library.
Before the change, the coupling caused conflicting pointers to ui components causing Angular apps rendered on browsers to get confused whether to render the angular compnent of the Lit component.
By decoupling the library, this change effectively omits the custom components defined for the Lit app to creep into the Angular app's compilation.
