+++ angular/projects/a2a-chat-canvas-demo/src/a2ui-catalog/google-map.ts
@@ -15,7 +15,7 @@
  */
 
 import { DynamicComponent } from '@a2ui/angular';
-import { v0_8 } from '@a2ui/web-lib';
+import * as v0_8 from '@a2ui/web-lib/0.8';
 import { CommonModule } from '@angular/common';
 import { Component, computed, input } from '@angular/core';
 import { GoogleMapsModule } from '@angular/google-maps';
