+++ angular/projects/a2a-chat-canvas/src/lib/services/canvas-service.ts
@@ -14,7 +14,7 @@
  limitations under the License.
  */
 
-import { v0_8 } from '@a2ui/web-lib';
+import * as v0_8 from '@a2ui/web-lib/0.8';
 import { Injectable, signal } from '@angular/core';
 
 /**
