+++ angular/projects/lib/src/lib/rendering/catalog.ts
@@ -16,7 +16,7 @@
 
 import { Binding, InjectionToken, Type } from '@angular/core';
 import { DynamicComponent } from './dynamic-component';
-import { v0_8 } from '@a2ui/web-lib';
+import * as v0_8 from '@a2ui/web-lib/0.8';
 
 export type CatalogLoader = () =>
   | Promise<Type<DynamicComponent<any>>>
