+++ angular/projects/lib/src/lib/rendering/dynamic-component.ts
@@ -14,7 +14,7 @@
  limitations under the License.
  */
 
-import { v0_8 } from '@a2ui/web-lib';
+import * as v0_8 from '@a2ui/web-lib/0.8';
 import { Directive, inject, input } from '@angular/core';
 import { ModelProcessor } from '../data';
 import { Theme } from './theming';
