+++ angular/projects/restaurant/src/app/client.ts
@@ -15,7 +15,7 @@
  */
 
 import { A2AServerPayload, ModelProcessor } from '@a2ui/angular';
-import { v0_8 } from '@a2ui/web-lib';
+import * as v0_8 from '@a2ui/web-lib/0.8';
 import { inject, Injectable, signal } from '@angular/core';
 
 @Injectable({ providedIn: 'root' })
