+++ angular/projects/rizzcharts/src/components/chat/message/message-content/message-content.ts
@@ -16,7 +16,7 @@
 
 import { DataPart, TextPart } from '@a2a-js/sdk';
 import { Surface } from '@a2ui/angular';
-import { v0_8 } from '@a2ui/web-lib';
+import * as v0_8 from '@a2ui/web-lib/0.8';
 import { Component, computed, inject, input, Signal } from '@angular/core';
 import { ChatService } from '@rizzcharts/services/chat_service';
 import { UiMessageContent } from '@rizzcharts/types/ui_message';
