+++ angular/projects/rizzcharts/src/services/chat_service.ts
@@ -16,7 +16,7 @@
 
 import { Artifact, DataPart, Message, Part, Task, TextPart } from '@a2a-js/sdk';
 import { ModelProcessor } from '@a2ui/angular';
-import { v0_8 } from '@a2ui/web-lib';
+import * as v0_8 from '@a2ui/web-lib/0.8';
 import { inject, Injectable, signal } from '@angular/core';
 import { UiMessage, UiMessageContent } from '@rizzcharts/types/ui_message';
 import { v4 as uuid } from 'uuid';
