+++ web/lib/package.json
@@ -10,6 +10,10 @@
       "types": "./dist/src/index.d.ts",
       "default": "./dist/src/index.js"
     },
+    "./0.8": {
+      "types": "./dist/src/0.8/core.d.ts",
+      "default": "./dist/src/0.8/core.js"
+    },
     "./ui": {
       "types": "./dist/src/0.8/ui/ui.d.ts",
       "default": "./dist/src/0.8/ui/ui.js"
