+++ web/lib/src/0.8/index.ts
@@ -14,23 +14,5 @@
  limitations under the License.
  */
 
-export * as Events from "./events/events.js";
-export * as Types from "./types/types.js";
-export * as Primitives from "./types/primitives.js";
-export * as Styles from "./styles/index.js";
+export * from "./core.js";
 export * as UI from "./ui/ui.js";
-import * as Guards from "./data/guards.js";
-
-import { create as createSignalA2UIModelProcessor } from "./data/signal-model-processor.js";
-import { A2UIModelProcessor } from "./data/model-processor.js";
-import A2UIClientEventMessage from "./schemas/server_to_client_with_standard_catalog.json" with { type: "json" };
-
-export const Data = {
-  createSignalA2UIModelProcessor,
-  A2UIModelProcessor,
-  Guards,
-};
-
-export const Schemas = {
-  A2UIClientEventMessage,
-};
